Struct nautilus_adapters::databento::loader::DatabentoDataLoader

pub struct DatabentoDataLoader { /* private fields */ }

Provides a Nautilus data loader for Databento Binary Encoding (DBN) format data.

Supported schemas:

MBO -> OrderBookDelta

MBP_1 -> (QuoteTick, Option<TradeTick>)

MBP_10 -> OrderBookDepth10

TBBO -> (QuoteTick, TradeTick)

TRADES -> TradeTick

OHLCV_1S -> Bar

OHLCV_1M -> Bar

OHLCV_1H -> Bar

OHLCV_1D -> Bar

DEFINITION -> Instrument

IMBALANCE -> DatabentoImbalance

STATISTICS -> DatabentoStatistics

References

https://databento.com/docs/knowledge-base/new-users/dbn-encoding

Implementations

impl DatabentoDataLoader

pub fn new(path: Option<PathBuf>) -> Result<Self>

pub fn load_publishers(&mut self, path: PathBuf) -> Result<()>

Load the publishers data from the file at the given path.

pub fn get_publishers(&self) -> &IndexMap<u16, DatabentoPublisher>

Return the internal Databento publishers currently held by the loader.

pub fn get_dataset_for_venue(&self, venue: &Venue) -> Option<&Dataset>

pub fn get_venue_for_publisher( &self, publisher_id: PublisherId ) -> Option<&Venue>

pub fn schema_from_file(&self, path: PathBuf) -> Result<Option<String>>

pub fn read_definition_records( &mut self, path: PathBuf ) -> Result<impl Iterator<Item = Result<InstrumentAny>> + '_>

pub fn read_records<T>( &self, path: PathBuf, instrument_id: Option<InstrumentId>, include_trades: bool ) -> Result<impl Iterator<Item = Result<(Option<Data>, Option<Data>)>> + '_>
where T: Record + HasRType + 'static,

pub fn read_imbalance_records<T>( &self, path: PathBuf, instrument_id: Option<InstrumentId> ) -> Result<impl Iterator<Item = Result<DatabentoImbalance>> + '_>
where T: Record + HasRType + 'static,

pub fn read_statistics_records<T>( &self, path: PathBuf, instrument_id: Option<InstrumentId> ) -> Result<impl Iterator<Item = Result<DatabentoStatistics>> + '_>
where T: Record + HasRType + 'static,