Struct nautilus_adapters::databento::loader::DatabentoDataLoader

pub struct DatabentoDataLoader { /* private fields */ }

Provides a Nautilus data loader for Databento Binary Encoding (DBN) format data.

Supported schemas:

MBO -> OrderBookDelta

MBP_1 -> QuoteTick | TradeTick

MBP_10 -> OrderBookDepth10

TBBO -> QuoteTick | TradeTick

TRADES -> TradeTick

OHLCV_1S -> Bar

OHLCV_1M -> Bar

OHLCV_1H -> Bar

OHLCV_1D -> Bar

DEFINITION -> Instrument

IMBALANCE -> DatabentoImbalance

STATISTICS -> DatabentoStatistics

For the loader to work correctly, you must first either:

Load Databento instrument definitions from a DBN file using load_instruments(...)

Manually add Nautilus instrument objects through add_instruments(...)

Warnings

The following Databento instrument classes are not supported:

FUTURE_SPREAD

OPTION_SPEAD

MIXED_SPREAD

FX_SPOT

References

https://docs.databento.com/knowledge-base/new-users/dbn-encoding

Implementations

impl DatabentoDataLoader

pub fn new(path: Option<PathBuf>) -> Result<Self>

pub fn load_publishers(&mut self, path: PathBuf) -> Result<()>

Load the publishers data from the file at the given path.

pub fn get_publishers(&self) -> &IndexMap<u16, DatabentoPublisher>

Return the internal Databento publishers currently held by the loader.

pub fn get_dataset_for_venue(&self, venue: &Venue) -> Option<&Dataset>

pub fn get_nautilus_instrument_id_for_record( &self, record: &RecordRef<'_>, metadata: &Metadata ) -> Result<InstrumentId>

pub fn schema_from_file(&self, path: PathBuf) -> Result<Option<String>>

pub fn read_records<T>( &self, path: PathBuf, instrument_id: Option<InstrumentId> ) -> Result<impl Iterator<Item = Result<(Data, Option<Data>)>> + '_>
where T: Record + HasRType + 'static,

pub fn read_definition_records( &self, path: PathBuf ) -> Result<impl Iterator<Item = Result<Box<dyn Instrument>>> + '_>

impl DatabentoDataLoader

pub fn py_new(path: Option<String>) -> PyResult<Self>

pub fn py_schema_for_file(&self, path: String) -> PyResult<Option<String>>

pub fn py_load_instruments( &self, py: Python<'_>, path: String ) -> PyResult<PyObject>

pub fn py_load_order_book_deltas( &self, path: String, instrument_id: Option<InstrumentId> ) -> PyResult<Vec<OrderBookDelta>>

pub fn py_load_order_book_depth10( &self, path: String, instrument_id: Option<InstrumentId> ) -> PyResult<Vec<OrderBookDepth10>>

pub fn py_load_quote_ticks( &self, path: String, instrument_id: Option<InstrumentId> ) -> PyResult<Vec<QuoteTick>>

pub fn py_load_tbbo_trade_ticks( &self, path: String, instrument_id: Option<InstrumentId> ) -> PyResult<Vec<TradeTick>>

pub fn py_load_trade_ticks( &self, path: String, instrument_id: Option<InstrumentId> ) -> PyResult<Vec<TradeTick>>

pub fn py_load_bars( &self, path: String, instrument_id: Option<InstrumentId> ) -> PyResult<Vec<Bar>>